pyproject.toml
[project]
name = "fucknround"
version = "0.1.0"
description = "Hopefully we get something of a data pipeline here soon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "requests",
    "rnet>=2.2.11",
]

[tool.setuptools]
package-dir = {"" = "src"}
